requirements.txt
pandas
numpy
lxml
mysql-connector-python
flask
mysql-connector-python
pandas
flask-cors  # Đã thêm